Handling errors that can occur during http request call
Introduce a check in the request() call. The check will verify if
there is an error. If there is an error, an error message is displayed on the console.
If there isn't an error the excution of the request call continue and the weather data is displayed on the console.