Use the callback pattern to improve the weather_app application
create a geoCode function that takes in the location address and a callback function
Inside the geoCode function, we call the request function to call the geocode mapbox api
By creating a function for this feature we want the code to be reusable and maintainable.
It is also easy to do one this befor or after something else.